Changes made as requested.
I actually added appsettings.json as it was not there and not mentioned in the docs. I had to change the default value of the connection string name to piranha. And it works in my case. I'm only adding this so that others donot face the same difficulty I did.